gain_admin_access - Medium-Hard/flask/requirements.txt
flask
pyjwt
cryptography
flask_cors